+++ demo/dataset/robotcar/lidar_cloud/main.cpp
@@ -1,5 +1,6 @@
 #include <opencv2/imgproc.hpp>
-#include "dataset/RobotcarReader.h"
+
+#include "dataset/robotcar_reader.h"
 #include "util.h"
 
 DEFINE_string(models_dir, "data/models/robotcar",
