+++ include/camera/camera_model_scaramuzza.h
