--- test/test_util.cpp
@@ -1,12 +1,0 @@
-#include <glog/logging.h>
-#include <gtest/gtest.h>
-
-#include "util.h"
-
-int main(int argc, char **argv) {
-  ::testing::InitGoogleTest(&argc, argv);
-  gflags::ParseCommandLineFlags(&argc, &argv, true);
-  google::InitGoogleLogging(argv[0]);
-
-  return RUN_ALL_TESTS();
-}
